requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9
alembic>=1.13.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
